requirements.txt
discord.py
requests
datetime
humanfriendly